+++ lib/xmi/uml/owned_attribute.rb
@@ -35,9 +35,10 @@ module Xmi
         map_attribute "aggregation", to: :aggregation
 
         map_element "type", to: :uml_type
-        map_element "upperValue", to: :upper_value,
-                                  polymorphic: VALUE_SPECIFICATION_POLYMORPHIC_MAP
+        # Sparx EA emits lowerValue before upperValue.
         map_element "lowerValue", to: :lower_value,
+                                  polymorphic: VALUE_SPECIFICATION_POLYMORPHIC_MAP
+        map_element "upperValue", to: :upper_value,
                                   polymorphic: VALUE_SPECIFICATION_POLYMORPHIC_MAP
         map_element "defaultValue", to: :default_value,
                                     polymorphic: VALUE_SPECIFICATION_POLYMORPHIC_MAP